backend/requirements.txt
flask==3.1.*
flask-cors==5.*
yt-dlp>=2024.0
fastai==2.8.6
plum-dispatch==2.6.1
Pillow>=10.0
torch>=2.0
torchvision>=0.15
gunicorn==22.*
ffmpeg-python==0.2.0
ipython==9.*